function showAllCharacters and function searchButton
