+++ src/index.html
@@ -10,7 +10,35 @@
 </header>
 <div class="buscador">
   <input type="text" placeholder="Buscar" class="search">
-  <button type="submit"><img src="./img/lupa.png" alt="" class="btnLupa" id="lupa"></button>     
+  
+  <div class="rdContainer">
+    <div class="rdBtns">
+    <input type="radio" id="rdOptions" name="rdOpt" value="nombre">
+    <label for="nombre">Nombre</label>
+    </div>
+
+    <div class="rdBtns">
+    <input type="radio" id="rdOptions" name="rdOpt" value="especie">
+    <label for="especie">Especie</label>
+    </div>
+
+    <div class="rdBtns">
+    <input type="radio" id="rdOptions" name="rdOpt" value="status">
+    <label for="status">Status</label>
+    </div>
+
+    <div class="rdBtns">
+    <input type="radio" id="rdOptions" name="rdOpt" value="locaciones">
+    <label for="locacion">Locaciones</label>
+    </div>
+
+    <div class="rdBtns">
+    <input type="radio" id="rdOptions" name="rdOpt" value="episodios">
+    <label for="episodios">Episodios</label>
+    </div>
+  </div>
+
+  <button type="button" id="searchBtn"><img src="./img/lupa.png" alt="" class="btnLupa" id="lupa"></button>
 </div>
   <body class="containerPrinc" width="100%">
     <main>
